LPS-140437 search-experiences-test:fix wrong class name pattern and re-group the test cases.
https://issues.liferay.com/browse/LPS-140437
