+++ modules/dxp/apps/search-experiences/search-experiences-test/src/testIntegration/java/com/liferay/search/experiences/internal/blueprint/condition/test/ConditionRangeSXPBlueprintTest.java
@@ -40,8 +40,7 @@ import org.junit.runner.RunWith;
  * @author Wade Cao
  */
 @RunWith(Arquillian.class)
-public class SXPBlueprintBoostContentsInCategoryForAPeriodOfTimeTest
-	extends BaseSXPBlueprintsTestCase {
+public class ConditionRangeSXPBlueprintTest extends BaseSXPBlueprintsTestCase {
 
 	@ClassRule
 	@Rule
@@ -54,51 +53,95 @@ public class SXPBlueprintBoostContentsInCategoryForAPeriodOfTimeTest
 	public void setUp() throws Exception {
 		super.setUp();
 
-		addJournalArticle(group.getGroupId(), "Coca Cola", "cola cola");
-
-		_assetCategory = addAssetCategory(
-			"Promoted", group,
-			SXPBlueprintTestUtil.addGroupUser(group, "Custmers"));
-
-		serviceContext.setAssetCategoryIds(
-			new long[] {_assetCategory.getCategoryId()});
-
-		addJournalArticle(group.getGroupId(), "Pepsi Cola", "");
-
-		_startDate = new Date(System.currentTimeMillis());
-		_endDate = _getNextDay();
-
 		setUpSXPBlueprint(getClass());
-	}
-
-	@Test
-	public void testSearch() throws Exception {
-		updateSXPBlueprint(getEmptyConfigurationJSONString());
-
-		assertSearchIgnoreRelevance("[coca cola, pepsi cola]", "cola");
 	}
 
 	@Ignore
 	@Test
-	public void testSearchWithInRangeCondition() throws Exception {
-		updateSXPBlueprint(_getConfigurationJSONString());
+	public void testSearchAPeriodOfTimeWithConditionRange() throws Exception {
+		addJournalArticle(group.getGroupId(), "Coca Cola", "cola cola");
+
+		AssetCategory assetCategory = addAssetCategory(
+			"Promoted", group,
+			SXPBlueprintTestUtil.addGroupUser(group, "Custmers"));
+
+		serviceContext.setAssetCategoryIds(
+			new long[] {assetCategory.getCategoryId()});
+
+		addJournalArticle(group.getGroupId(), "Pepsi Cola", "");
+
+		updateSXPBlueprint(
+			_getConfigurationJSONString(
+				assetCategory, new Date(System.currentTimeMillis()),
+				_getNextDay()));
 
 		assertSearch("[pepsi cola, coca cola]", "cola");
 	}
 
-	private String _getConfigurationJSONString() {
+	@Ignore
+	@Test
+	public void testSearchNewUserAccountWithConditionRange() throws Exception {
+		updateSXPBlueprint(
+			StringUtil.replace(
+				getConfigurationJSONString(
+					getClass(), testName.getMethodName()),
+				"${configuration.asset_category_id}",
+				String.valueOf(
+					_addAssetCategoryJournalArticles().getCategoryId())));
+
+		assertSearch(
+			"[company policies for new recruits, company policies for all " +
+				"employees recruits]",
+			"policies");
+	}
+
+	@Test
+	public void testSearchNewUserAccountWithoutRangeCondition()
+		throws Exception {
+
+		_addAssetCategoryJournalArticles();
+
+		updateSXPBlueprint(getEmptyConfigurationJSONString());
+
+		assertSearchIgnoreRelevance(
+			"[company policies for all employees recruits, company policies " +
+				"for new recruits]",
+			"policies");
+	}
+
+	private AssetCategory _addAssetCategoryJournalArticles() throws Exception {
+		AssetCategory assetCategory = addAssetCategory(
+			"For New Recruits", group,
+			SXPBlueprintTestUtil.addGroupUser(group, "Employee"));
+
+		addJournalArticle(
+			group.getGroupId(), "Company policies for All Employees Recruits",
+			"policies policies");
+
+		serviceContext.setAssetCategoryIds(
+			new long[] {assetCategory.getCategoryId()});
+
+		addJournalArticle(
+			group.getGroupId(), "Company Policies for New Recruits", "");
+
+		return assetCategory;
+	}
+
+	private String _getConfigurationJSONString(
+		AssetCategory assetCategory, Date startDate, Date endDate) {
+
 		String configurationJSON = StringUtil.replace(
-			getConfigurationJSONString(getClass()),
+			getConfigurationJSONString(getClass(), testName.getMethodName()),
 			"${configuration.asset_category_id}",
-			String.valueOf(_assetCategory.getCategoryId()));
+			String.valueOf(assetCategory.getCategoryId()));
 
 		configurationJSON = StringUtil.replace(
 			configurationJSON, "${configuration.start_date}",
-			DateUtil.getDate(_startDate, "yyyyMMdd", LocaleUtil.US));
+			DateUtil.getDate(startDate, "yyyyMMdd", LocaleUtil.US));
 
 		return StringUtil.replace(
 			configurationJSON, "${configuration.end_date}",
-			DateUtil.getDate(_endDate, "yyyyMMdd", LocaleUtil.US));
+			DateUtil.getDate(endDate, "yyyyMMdd", LocaleUtil.US));
 	}
 
 	private Date _getNextDay() {
@@ -108,9 +151,5 @@ public class SXPBlueprintBoostContentsInCategoryForAPeriodOfTimeTest
 
 		return cal.getTime();
 	}
-
-	private AssetCategory _assetCategory;
-	private Date _endDate;
-	private Date _startDate;
 
 }
