--- modules/dxp/apps/search-experiences/search-experiences-test/src/testIntegration/java/com/liferay/search/experiences/internal/blueprint/condition/test/SXPBlueprintBoostWebContentByKeywordMatchTest.java
@@ -1,89 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * The contents of this file are subject to the terms of the Liferay Enterprise
- * Subscription License ("License"). You may not use this file except in
- * compliance with the License. You can obtain a copy of the License by
- * contacting Liferay, Inc. See the License for the specific language governing
- * permissions and limitations under the License, including but not limited to
- * distribution rights of the Software.
- *
- *
- *
- */
-
-package com.liferay.search.experiences.internal.blueprint.condition.test;
-
-import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.journal.model.JournalArticle;
-import com.liferay.portal.kernel.test.rule.AggregateTestRule;
-import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
-import com.liferay.portal.test.rule.PermissionCheckerMethodTestRule;
-import com.liferay.search.experiences.internal.blueprint.test.BaseSXPBlueprintsTestCase;
-
-import org.junit.Before;
-import org.junit.ClassRule;
-import org.junit.Ignore;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-/**
- * @author Wade Cao
- */
-@RunWith(Arquillian.class)
-public class SXPBlueprintBoostWebContentByKeywordMatchTest
-	extends BaseSXPBlueprintsTestCase {
-
-	@ClassRule
-	@Rule
-	public static final AggregateTestRule aggregateTestRule =
-		new AggregateTestRule(
-			new LiferayIntegrationTestRule(),
-			PermissionCheckerMethodTestRule.INSTANCE);
-
-	@Before
-	public void setUp() throws Exception {
-		super.setUp();
-
-		addJournalArticle(group.getGroupId(), "Coca Cola", "cola cola");
-
-		_journalArticle = addJournalArticle(
-			group.getGroupId(), "Pepsi Cola", "");
-
-		setUpSXPBlueprint(getClass());
-	}
-
-	@Ignore
-	@Test
-	public void testConditionWithContains() throws Exception {
-		updateSXPBlueprint(_getConfigurationJSONString());
-
-		assertSearch("[pepsi cola, coca cola]", "cola");
-	}
-
-	@Ignore
-	@Test
-	public void testConditionWithNotContains() throws Exception {
-		updateSXPBlueprint(_getConfigurationJSONString());
-
-		assertSearch("[coca cola, pepsi cola]", "cola");
-	}
-
-	@Test
-	public void testSearch() throws Exception {
-		updateSXPBlueprint(getEmptyConfigurationJSONString());
-
-		assertSearchIgnoreRelevance("[coca cola, pepsi cola]", "cola");
-	}
-
-	private String _getConfigurationJSONString() {
-		return StringUtil.replace(
-			getConfigurationJSONString(getClass(), testName.getMethodName()),
-			"${articleId}", _journalArticle.getArticleId());
-	}
-
-	private JournalArticle _journalArticle;
-
-}
