+++ modules/dxp/apps/search-experiences/search-experiences-test/src/testIntegration/java/com/liferay/search/experiences/internal/blueprint/test/BaseSXPBlueprintsTestCase.java
@@ -229,6 +229,7 @@ public abstract class BaseSXPBlueprintsTestCase {
 				searchContext.setAttribute(
 					"search.experiences.blueprint.id",
 					_sxpBlueprint.getSXPBlueprintId());
+				searchContext.setTimeZone(user.getTimeZone());
 				searchContext.setUserId(user.getUserId());
 			}
 		).build();
