+++ modules/dxp/apps/search-experiences/search-experiences-test/src/testIntegration/java/com/liferay/search/experiences/internal/blueprint/test/TextMatchOverMultipleFieldsSXPBlueprintTest.java
@@ -41,7 +41,7 @@ import org.junit.runner.RunWith;
  * @author Wade Cao
  */
 @RunWith(Arquillian.class)
-public class SXPBlueprintTextMatchOverMultipleFieldsTest
+public class TextMatchOverMultipleFieldsSXPBlueprintTest
 	extends BaseSXPBlueprintsTestCase {
 
 	@ClassRule
