+++ modules/dxp/apps/search-experiences/search-experiences-test/src/testIntegration/resources/com/liferay/search/experiences/internal/blueprint/condition/test/dependencies/ConditionContainsSXPBlueprintTest.testSearchWithAssetCategory.json
@@ -1,4 +1,10 @@
 {
+	"general": {
+		"searchableAssetTypes": [
+			"com.liferay.wiki.model.WikiPage",
+			"com.liferay.journal.model.JournalArticle"
+		]
+	},
 	"queryConfiguration": {
 		"applyIndexerClauses": false,
 		"queryEntries": [
@@ -19,8 +25,8 @@
 				],
 				"condition": {
 					"contains": {
-						"parameterName": "${keywords}",
-						"value": "cola"
+						"parameterName": "keywords",
+						"value": "${configuration.keywords}"
 					}
 				},
 				"enabled": true
