+++ modules/dxp/apps/search-experiences/search-experiences-test/src/testIntegration/resources/com/liferay/search/experiences/internal/blueprint/condition/test/dependencies/ConditionContainsSXPBlueprintTest.testSearchWithConditionNotContains.json
@@ -1,0 +1,36 @@
+{
+	"general": {
+		"searchableAssetTypes": [
+			"com.liferay.wiki.model.WikiPage",
+			"com.liferay.journal.model.JournalArticle"
+		]
+	},
+	"queryConfiguration": {
+		"applyIndexerClauses": false,
+		"queryEntries": [
+			{
+				"clauses": [
+					{
+						"context": "query",
+						"occur": "should",
+						"query": {
+							"terms": {
+								"articleId_String_sortable": [
+									"${articleId}"
+								],
+								"boost": 100
+							}
+						}
+					}
+				],
+				"condition": {
+					"contains": {
+						"parameterName": "keywords",
+						"value": "cola"
+					}
+				},
+				"enabled": true
+			}
+		]
+	}
+}
