+++ modules/dxp/apps/search-experiences/search-experiences-test/src/testIntegration/resources/com/liferay/search/experiences/internal/blueprint/condition/test/dependencies/ConditionContainsSXPBlueprintTest.testSearchWithSegmentsEntry.json
@@ -1,4 +1,10 @@
 {
+	"general": {
+		"searchableAssetTypes": [
+			"com.liferay.wiki.model.WikiPage",
+			"com.liferay.journal.model.JournalArticle"
+		]
+	},
 	"queryConfiguration": {
 		"applyIndexerClauses": false,
 		"queryEntries": [
@@ -19,10 +25,8 @@
 				],
 				"condition": {
 					"contains": {
-						"parameterName": "${user.active_segment_entry_ids}",
-						"value": [
-							"${configuration.user_segment_ids}"
-						]
+						"parameterName": "user.active_segment_entry_ids",
+						"value": "${configuration.user_segment_ids}"
 					}
 				},
 				"enabled": true
