+++ modules/dxp/apps/search-experiences/search-experiences-test/src/testIntegration/resources/com/liferay/search/experiences/internal/blueprint/test/dependencies/BoostContentsSXPBlueprintTest.testSearchBoostContentsInCategory.json
@@ -1,6 +1,12 @@
 {
+	"general": {
+		"searchableAssetTypes": [
+			"com.liferay.wiki.model.WikiPage",
+			"com.liferay.journal.model.JournalArticle"
+		]
+	},
 	"queryConfiguration": {
-		"applyIndexerClauses": false,
+		"applyIndexerClauses": true,
 		"queryEntries": [
 			{
 				"clauses": [
@@ -9,8 +15,8 @@
 						"occur": "should",
 						"query": {
 							"terms": {
-								"articleId_String_sortable": [
-									"${articleId}"
+								"assetCategoryIds": [
+									"${configuration.asset_category_ids}"
 								],
 								"boost": 100
 							}
@@ -18,10 +24,6 @@
 					}
 				],
 				"condition": {
-					"contains": {
-						"parameterName": "${keywords}",
-						"value": "cola"
-					}
 				},
 				"enabled": true
 			}
