+++ modules/dxp/apps/search-experiences/search-experiences-test/src/testIntegration/resources/com/liferay/search/experiences/internal/blueprint/test/dependencies/LimitSearchSXPBlueprintTest.testSearchWithLimitSearchToTheseSites.json
@@ -16,8 +16,8 @@
 						"query": {
 							"terms": {
 								"scopeGroupId": [
-									"${configuration.scope_group_id_1}",
-									"${configuration.scope_group_id_2}"
+									"${configuration.value1}",
+									"${configuration.value2}"
 								]
 							}
 						}
