--- modules/dxp/apps/search-experiences/search-experiences-test/src/testIntegration/resources/com/liferay/search/experiences/internal/blueprint/test/dependencies/SXPBlueprintBoostContentsInCategoryTest.json
@@ -1,26 +1,0 @@
-{
-	"queryConfiguration": {
-		"applyIndexerClauses": true,
-		"queryEntries": [
-			{
-				"clauses": [
-					{
-						"context": "query",
-						"occur": "should",
-						"query": {
-							"terms": {
-								"assetCategoryIds": [
-									"${configuration.asset_category_ids}"
-								],
-								"boost": 100
-							}
-						}
-					}
-				],
-				"condition": {
-				},
-				"enabled": true
-			}
-		]
-	}
-}
