+++ modules/dxp/apps/search-experiences/search-experiences-test/src/testIntegration/resources/com/liferay/search/experiences/internal/blueprint/test/dependencies/TextMatchOverMultipleFieldsSXPBlueprintTest.testSearchCrossFieldsWithOperator.json
